feat: finished migrating donation page
