+++ app/donation/page.tsx
@@ -1,0 +1,9 @@
+import { Donation } from './Donation';
+
+export const metadata = {
+	title: 'Donation',
+};
+
+export default function Page() {
+	return <Donation />;
+}
